requirements.txt
altgraph==0.17.4
beautifulsoup4==4.12.3
clipboard==0.0.4
keyboard==0.13.5
mouse==0.7.1
packaging==24.2
pefile==2023.2.7
psutil==6.1.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyperclip==1.9.0
python-osc==1.9.0
pywin32==308
pywin32-ctypes==0.2.3
setuptools==75.6.0
soupsieve==2.6
WinGrab==0.0.1a4
WMI==1.5.1
